Refactor fuel-asm lib parts into conv and op modules
